+++ tests/test_translate_document.py
@@ -157,7 +157,9 @@ def test_translate_document_low_level(
     assert status.ok and not status.done
 
     # Calling download() before document is ready will fail
-    with pytest.raises(deepl.DeepLException, match="Document not ready"):
+    with pytest.raises(
+        deepl.DocumentNotReadyException, match="Document not ready"
+    ):
         with open(output_document_path, "wb") as output_file:
             translator.translate_document_download(handle, output_file)